realtors/requirements.txt
Flask-Script
Flask-SQLAlchemy
alembic
psycopg2-binary
psycopg2
requests
celery
redis
